+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/service/DeleteMyMentorInfoUseCase.kt
@@ -1,0 +1,6 @@
+package team.themoment.gsmNetworking.domain.mentor.service
+
+interface DeleteMyMentorInfoUseCase {
+
+    fun deleteMyMentorInfo(authenticationId: Long)
+}
