+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/service/DeleteTempMentorByIdUseCase.kt
@@ -1,0 +1,6 @@
+package team.themoment.gsmNetworking.domain.mentor.service
+
+interface DeleteTempMentorByIdUseCase {
+
+    fun deleteTempMentorById(id: Long)
+}
