+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/service/MentorRegistrationUseCase.kt
@@ -1,0 +1,8 @@
+package team.themoment.gsmNetworking.domain.mentor.service
+
+import team.themoment.gsmNetworking.domain.mentor.dto.MentorRegistrationDto
+
+interface MentorRegistrationUseCase {
+
+    fun mentorRegistration(dto: MentorRegistrationDto, authenticationId: Long)
+}
