+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/service/ModifyMyMentorInfoUseCase.kt
@@ -1,0 +1,8 @@
+package team.themoment.gsmNetworking.domain.mentor.service
+
+import team.themoment.gsmNetworking.domain.mentor.dto.MentorUpdateInfoDto
+
+interface ModifyMyMentorInfoUseCase {
+
+    fun modifyMyMentorInfo(mentorUpdateInfoDto: MentorUpdateInfoDto)
+}
