+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/service/QueryAllTempMentorsUseCase.kt
@@ -1,0 +1,8 @@
+package team.themoment.gsmNetworking.domain.mentor.service
+
+import team.themoment.gsmNetworking.domain.mentor.dto.TempMentorInfoDto
+
+interface QueryAllTempMentorsUseCase {
+
+    fun queryAllTempMentors(): List<TempMentorInfoDto>
+}
