+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/service/QueryMyMentorInfoUseCase.kt
@@ -1,0 +1,8 @@
+package team.themoment.gsmNetworking.domain.mentor.service
+
+import team.themoment.gsmNetworking.domain.mentor.dto.MyMentorInfoDto
+
+interface QueryMyMentorInfoUseCase {
+
+    fun queryMyMentorInfo(authenticationId: Long): MyMentorInfoDto
+}
