+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/service/QueryTempMentorByIdUseCase.kt
@@ -1,0 +1,8 @@
+package team.themoment.gsmNetworking.domain.mentor.service
+
+import team.themoment.gsmNetworking.domain.mentor.dto.TempMentorInfoDto
+
+interface QueryTempMentorByIdUseCase {
+
+    fun queryTempMentorById(id: Long): TempMentorInfoDto
+}
