+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/service/QueryTempMentorsByNameUseCase.kt
@@ -1,0 +1,8 @@
+package team.themoment.gsmNetworking.domain.mentor.service
+
+import team.themoment.gsmNetworking.domain.mentor.dto.SearchTempMentorInfoDto
+
+interface QueryTempMentorsByNameUseCase {
+
+    fun queryTempMentorsByName(name: String): List<SearchTempMentorInfoDto>
+}
